--- challenges/02-bank/submissions/guilherme-portas/src/Customer.java
@@ -1,10 +1,0 @@
-public class Customer {
-    String firstName, lastName, CPF;
-    Account accounts;
-
-    public void customer(String firstName, String lastName, String CPF) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.CPF = CPF;
-    }
-}
