--- challenges/02-bank/submissions/pedro-dutra/src/Account.java
@@ -1,3 +1,0 @@
-public class Account {
-    
-}
