--- challenges/02-bank/submissions/pedro-dutra/src/IDPBank.java
@@ -1,19 +1,0 @@
-import java.util.Scanner;
-
-public class IDPBank {
- 
-    public static void main(String[] args){
-        Scanner scanner = new Scanner(System.in);
-        System.out.println(""
-                            + "Seja bem vindo ao IDP Bank! Escolha uma das opções abaixo: \n"
-                            + "1. Abrir conta \n"
-                            + "2. Consultar cliente \n"
-                            + "3. Depositar \n"
-                            + "4. Sacar \n"
-                            + "5. Sair \n \n"
-                            + "Digite o número correspondente a opção desejada: ");
-    
-    String opcao = scanner.nextLine();
-    
-    }
-}
