+++ challenges/bank02/submissions/guilherme-portas/src/Customer.java
@@ -1,0 +1,10 @@
+public class Customer {
+    String firstName, lastName, CPF;
+    Account accounts;
+
+    public void customer(String firstName, String lastName, String CPF) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.CPF = CPF;
+    }
+}
