+++ challenges/bank02/submissions/pedro-dutra/src/Account.java
@@ -1,0 +1,3 @@
+public class Account {
+    
+}
