+++ challenges/bank02/submissions/pedro-pignata/src/openAccount.java
@@ -1,0 +1,4 @@
+
+public class openAccount {
+
+}
